+++ lib/src/StkTable/index.d.ts
@@ -1,3 +1,5 @@
 export { default as StkTable } from './StkTable.vue';
 export { tableSort, insertToOrderedArray, strCompare, binarySearch } from './utils';
 export type { StkTableColumn } from './types/index';
+export { useAreaSelection } from './features/index';
+export { registerFeature } from './registerFeature';